Merged PR 88411: Fix html self closing issues
**Issue**
Q&A elements are rendered in a wrong way.

**Reason**
Some div elements are marked with self-closing.
According to the HTML5 spec, tags that cannot have any contents (known as void elements) can be self-closing*. This includes the following tags:

**Resolution**
_Only void elements can be self closing. from HTML5 spec, void elements are:_
area, base, br, col, embed, hr, img, input, link, meta, param, source, track, wbr

**Copied from HTML5 SPEC:**
Tags are used to delimit the start and end of elements in the markup. Raw text, escapable raw text, and normal elements have a start tag to indicate where they begin, and an end tag to indicate where they end. The start and end tags of certain normal elements can be omitted, as described below in the section on optional tags. Those that cannot be omitted must not be omitted. Void elements only have a start tag; end tags must not be specified for void elements. Foreign elements must either have a start tag and an end tag, or a start tag that is marked as self-closing, in which case they must not have an end tag.

**References:**
https://html.spec.whatwg.org/multipage/syntax.html#void-elements
https://stackoverflow.com/questions/3558119/are-non-void-self-closing-tags-valid-in-html5

Related work items: #365457