Merged PR 108503: Avoid sending another /load to the FE if one was already sent
A client has been reporting "content is not available" issues for some time.
We notice in their logs that several load requests are sent to the front end with only 66 millisecond between them.
We believe that this is causing a timing issue in the front end and so we add a throttle on the load requests.

## Notes for reviewers:
I am not sure if there can be any valid use case for multiple load requests in short intervals.
Can it negatively affect valid uses cases of bootstrap -> embed -> potentially reload flows?