Merged PR 136313: 2.16.3 global this
Between versions 2.15.1 and 2.16.1 Webpack version was updated which caused the compiled JS code to use the window object instead of the global `this` object.

`window` object is not defined in Node.JS so when I tried to update the version used by the front end E2E tests I got the following error:
It might also be the reason why the travis build fails the `npm install` in Github.

![e2e.PNG](https://powerbi.visualstudio.com/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/3e42401e-931d-4a96-b4b9-3533d1cf56ce/pullRequests/136311/attachments/e2e.PNG)