Merged PR 139689: Reduce number of eslint errors
There are currently about 2000 eslint warnings when running gulp build which slows the build time.
This PR reduces that number to about 60 by doing the following things;
- Run the automatic eslint fix
- Ignore the test folder
- Ignore several noisy rules
- Fix imports

Please let me know of any non-trivial change in this PR so I can revert it and make this PR simpler and safer.
I recommend reviewing this with `Show and diff white space` unchecked