Merged PR 275402: New API setSdkInfo in service + version bump 2.21.1
sdkType tells us which SDK wrapper performs the actions, e.g: powerbi-client-react, powerbi-client-angular...

To update it we should pass the type in the header 'x-sdk-type` which is populated in the service constructor. Therefore, each wrapper needs to pass the correct sdkType when initiating the service (see example: [Pull Request 274150](https://dev.azure.com/powerbi/Embedded/_git/powerbi-client-react/pullrequest/274150)).

**What is the problem?** if user injects a new powerbi service to the wrapper and doesn't use the default service, the sdkType won't be populated.
As a result, I created new API in the service, so we could update in each wrapper the sdk type after the service is created.

In addition, create new header `x-sdk-wrapper-version`.